+++ lib/minimap2/mmpriv.h
@@ -53,7 +53,7 @@ typedef struct {
 
 double cputime(void);
 double realtime(void);
-long peakrss(void);
+long long peakrss(void);
 
 void radix_sort_128x(mm128_t *beg, mm128_t *end);
 void radix_sort_64(uint64_t *beg, uint64_t *end);
